refactor: rename payment related files
- Rename payment_repository.py to payment.py
- Rename payment_service.py to payment.py
